Move `enrich` package under `outbus`
