+++ server/src/main/java/io/spine/server/outbus/enrich/EventEnricher.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.event.enrich;
+package io.spine.server.outbus.enrich;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Function;
