+++ server/src/test/java/io/spine/server/event/EventBusShould.java
@@ -31,8 +31,8 @@ import io.spine.core.Events;
 import io.spine.core.Subscribe;
 import io.spine.server.BoundedContext;
 import io.spine.server.bus.EnvelopeValidator;
-import io.spine.server.event.enrich.EventEnricher;
 import io.spine.server.event.given.EventBusTestEnv.GivenEvent;
+import io.spine.server.outbus.enrich.EventEnricher;
 import io.spine.server.storage.StorageFactory;
 import io.spine.test.event.ProjectCreated;
 import io.spine.test.event.ProjectId;
