+++ server/src/test/java/io/spine/server/outbus/enrich/EventEnricherBuilderShould.java
@@ -18,15 +18,15 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.event.enrich;
+package io.spine.server.outbus.enrich;
 
 import com.google.common.base.Function;
 import com.google.protobuf.StringValue;
 import com.google.protobuf.Timestamp;
 import com.google.protobuf.util.Timestamps;
 import io.spine.core.UserId;
-import io.spine.server.event.enrich.EventEnricher.SameTransition;
-import io.spine.server.event.enrich.given.EventEnricherBuilderTestEnv.Enrichment;
+import io.spine.server.outbus.enrich.EventEnricher.SameTransition;
+import io.spine.server.outbus.enrich.given.EventEnricherBuilderTestEnv.Enrichment;
 import io.spine.test.Tests;
 import io.spine.test.event.ProjectId;
 import org.junit.Before;
