+++ server/src/test/java/io/spine/server/outbus/enrich/EventEnricherShould.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.event.enrich;
+package io.spine.server.outbus.enrich;
 
 import com.google.common.base.Function;
 import com.google.common.base.Optional;
@@ -32,8 +32,8 @@ import io.spine.server.BoundedContext;
 import io.spine.server.command.TestEventFactory;
 import io.spine.server.event.EventBus;
 import io.spine.server.event.EventSubscriber;
-import io.spine.server.event.enrich.given.EventEnricherTestEnv.GivenEvent;
-import io.spine.server.event.enrich.given.EventEnricherTestEnv.GivenEventMessage;
+import io.spine.server.outbus.enrich.given.EventEnricherTestEnv.GivenEvent;
+import io.spine.server.outbus.enrich.given.EventEnricherTestEnv.GivenEventMessage;
 import io.spine.test.event.ProjectCompleted;
 import io.spine.test.event.ProjectCreated;
 import io.spine.test.event.ProjectCreatedDynamicallyConfiguredEnrichment;
@@ -51,11 +51,11 @@ import static io.spine.Identifier.newUuid;
 import static io.spine.core.Enrichments.getEnrichment;
 import static io.spine.protobuf.TypeConverter.toMessage;
 import static io.spine.server.command.TestEventFactory.newInstance;
-import static io.spine.server.event.enrich.given.EventEnricherTestEnv.Enrichment.GetProjectMaxMemberCount;
-import static io.spine.server.event.enrich.given.EventEnricherTestEnv.Enrichment.GetProjectName;
-import static io.spine.server.event.enrich.given.EventEnricherTestEnv.Enrichment.GetProjectOwnerId;
-import static io.spine.server.event.enrich.given.EventEnricherTestEnv.Enrichment.newEventEnricher;
-import static io.spine.server.event.enrich.given.EventEnricherTestEnv.GivenEvent.projectStarted;
+import static io.spine.server.outbus.enrich.given.EventEnricherTestEnv.Enrichment.GetProjectMaxMemberCount;
+import static io.spine.server.outbus.enrich.given.EventEnricherTestEnv.Enrichment.GetProjectName;
+import static io.spine.server.outbus.enrich.given.EventEnricherTestEnv.Enrichment.GetProjectOwnerId;
+import static io.spine.server.outbus.enrich.given.EventEnricherTestEnv.Enrichment.newEventEnricher;
+import static io.spine.server.outbus.enrich.given.EventEnricherTestEnv.GivenEvent.projectStarted;
 import static io.spine.testdata.TestBoundedContextFactory.MultiTenant.newBoundedContext;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
