+++ server/src/test/java/io/spine/server/outbus/enrich/given/EventEnricherTestEnv.java
@@ -18,30 +18,175 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.event.enrich.given;
+package io.spine.server.outbus.enrich.given;
 
 import com.google.common.base.Function;
 import com.google.protobuf.Any;
+import com.google.protobuf.Message;
 import com.google.protobuf.Timestamp;
 import io.spine.core.CommandContext;
+import io.spine.core.Event;
 import io.spine.core.EventId;
 import io.spine.core.UserId;
 import io.spine.core.given.GivenUserId;
 import io.spine.people.PersonName;
-import io.spine.server.event.enrich.EventEnricher;
+import io.spine.server.command.TestEventFactory;
+import io.spine.server.outbus.enrich.EventEnricher;
+import io.spine.test.event.ProjectCompleted;
+import io.spine.test.event.ProjectCreated;
 import io.spine.test.event.ProjectId;
+import io.spine.test.event.ProjectStarred;
+import io.spine.test.event.ProjectStarted;
+import io.spine.test.event.user.permission.PermissionGrantedEvent;
+import io.spine.test.event.user.permission.PermissionRevokedEvent;
+import io.spine.test.event.user.sharing.SharingRequestApproved;
 import io.spine.time.ZoneId;
 import io.spine.time.ZoneOffset;
 
 import javax.annotation.Nullable;
 
+import static io.spine.Identifier.newUuid;
+import static io.spine.protobuf.AnyPacker.pack;
+
 /**
  * @author Alexander Yevsyukov
  */
-public class EventEnricherBuilderTestEnv {
+public class EventEnricherTestEnv {
 
-    private EventEnricherBuilderTestEnv() {
+    private EventEnricherTestEnv() {
         // Prevent instantiation of this utility class.
+    }
+
+    static ProjectId newProjectId() {
+        final String uuid = newUuid();
+        return ProjectId.newBuilder()
+                        .setId(uuid)
+                        .build();
+    }
+
+    public static class GivenEventMessage {
+
+        private static final ProjectId PROJECT_ID = newProjectId();
+        private static final ProjectCreated PROJECT_CREATED = projectCreated(PROJECT_ID);
+        private static final ProjectStarted PROJECT_STARTED = projectStarted(PROJECT_ID);
+        private static final ProjectCompleted PROJECT_COMPLETED = projectCompleted(PROJECT_ID);
+        private static final ProjectStarred PROJECT_STARRED = projectStarred(PROJECT_ID);
+
+        private GivenEventMessage() {}
+
+        public static ProjectCreated projectCreated() {
+            return PROJECT_CREATED;
+        }
+
+        public static ProjectStarted projectStarted() {
+            return PROJECT_STARTED;
+        }
+
+        public static ProjectCompleted projectCompleted() {
+            return PROJECT_COMPLETED;
+        }
+
+        public static ProjectStarred projectStarred() {
+            return PROJECT_STARRED;
+        }
+
+        private static ProjectCreated projectCreated(ProjectId id) {
+            return ProjectCreated.newBuilder()
+                                 .setProjectId(id)
+                                 .build();
+        }
+
+        private static ProjectStarted projectStarted(ProjectId id) {
+            return ProjectStarted.newBuilder()
+                                 .setProjectId(id)
+                                 .build();
+        }
+
+        private static ProjectCompleted projectCompleted(ProjectId id) {
+            return ProjectCompleted.newBuilder()
+                                   .setProjectId(id)
+                                   .build();
+        }
+
+        private static ProjectStarred projectStarred(ProjectId id) {
+            return ProjectStarred.newBuilder()
+                                 .setProjectId(id)
+                                 .build();
+        }
+
+        static PermissionGrantedEvent permissionGranted() {
+            return PermissionGrantedEvent.newBuilder()
+                                         .setGranterUid(newUuid())
+                                         .setPermissionId("mock_permission")
+                                         .setUserUid(newUuid())
+                                         .build();
+        }
+
+        static PermissionRevokedEvent permissionRevoked() {
+            return PermissionRevokedEvent.newBuilder()
+                                         .setPermissionId("old_permission")
+                                         .setUserUid(newUuid())
+                                         .build();
+        }
+
+        static SharingRequestApproved sharingRequestApproved() {
+            return SharingRequestApproved.newBuilder()
+                                         .setUserUid(newUuid())
+                                         .build();
+        }
+    }
+
+    public static class GivenEvent {
+
+        private static final ProjectId PROJECT_ID = newProjectId();
+
+        private GivenEvent() {}
+
+        private static TestEventFactory eventFactory() {
+            final TestEventFactory result =
+                    TestEventFactory.newInstance(pack(PROJECT_ID), GivenEvent.class);
+            return result;
+        }
+
+        public static Event projectCreated() {
+            return projectCreated(PROJECT_ID);
+        }
+
+        public static Event projectStarted() {
+            final ProjectStarted msg = GivenEventMessage.projectStarted();
+            final Event event = eventFactory().createEvent(msg);
+            return event;
+        }
+
+        public static Event projectCreated(ProjectId projectId) {
+            final ProjectCreated msg = GivenEventMessage.projectCreated(projectId);
+            final Event event = eventFactory().createEvent(msg);
+            return event;
+        }
+
+        public static Event permissionGranted() {
+            final PermissionGrantedEvent message = GivenEventMessage.permissionGranted();
+            final Event permissionGranted = createGenericEvent(message);
+            return permissionGranted;
+        }
+
+        public static Event permissionRevoked() {
+            final PermissionRevokedEvent message = GivenEventMessage.permissionRevoked();
+            final Event permissionRevoked = createGenericEvent(message);
+            return permissionRevoked;
+        }
+
+        public static Event sharingRequestApproved() {
+            final SharingRequestApproved message = GivenEventMessage.sharingRequestApproved();
+            final Event sharingRequestApproved = createGenericEvent(message);
+            return sharingRequestApproved;
+        }
+
+        private static Event createGenericEvent(Message eventMessage) {
+            final Any wrappedMessage = pack(eventMessage);
+            final Event permissionRevoked = eventFactory().createEvent(wrappedMessage);
+            return permissionRevoked;
+        }
     }
 
     public static class Enrichment {
@@ -71,7 +216,7 @@ public class EventEnricherBuilderTestEnv {
                 if (id == null) {
                     return null;
                 }
-                final String name = "P-" + id.getId();
+                final String name = "prj_" + id.getId();
                 return name;
             }
         }
@@ -83,7 +228,18 @@ public class EventEnricherBuilderTestEnv {
                 if (id == null) {
                     return null;
                 }
-                return GivenUserId.of("PO-" + id.getId());
+                return GivenUserId.of("po_" + id.getId());
+            }
+        }
+
+        public static class GetProjectMaxMemberCount implements Function<ProjectId, Integer> {
+            @Nullable
+            @Override
+            public Integer apply(@Nullable ProjectId input) {
+                if (input == null) {
+                    return 0;
+                }
+                return input.hashCode();
             }
         }
 
@@ -148,8 +304,8 @@ public class EventEnricherBuilderTestEnv {
                     @Override
                     public PersonName apply(@Nullable String input) {
                         return input == null
-                                ? PersonName.getDefaultInstance()
-                                : PersonName.newBuilder().setFamilyName(input).build();
+                               ? PersonName.getDefaultInstance()
+                               : PersonName.newBuilder().setFamilyName(input).build();
                     }
                 };
